Update User.java
Please see that I change the datatype of mobile number from int to long

+++ Grocery1/src/main/demo/User.java
@@ -31,7 +31,7 @@ public class User {
 	
 	@NotNull
 	@Column(unique = true)
-	private int mobileNumber;
+	private Long mobileNumber;
 	
 	@Embedded
 	private Address address;
